+++ deploy/cloudflare/test/shims/cloudflare-workers.ts
@@ -8,3 +8,13 @@ export class DurableObject<Env> {
     this.env = env;
   }
 }
+
+export class WorkerEntrypoint<Env> {
+  protected readonly ctx: ExecutionContext;
+  protected readonly env: Env;
+
+  constructor(ctx: ExecutionContext, env: Env) {
+    this.ctx = ctx;
+    this.env = env;
+  }
+}